+++ src/poetry_plugin_export/command.py
@@ -1,13 +1,19 @@
 from __future__ import annotations
 
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 from cleo.helpers import option
+from packaging.utils import NormalizedName
 from packaging.utils import canonicalize_name
 from poetry.console.commands.group_command import GroupCommand
 from poetry.core.packages.dependency_group import MAIN_GROUP
 
 from poetry_plugin_export.exporter import Exporter
+
+
+if TYPE_CHECKING:
+    from collections.abc import Iterable
 
 
 class ExportCommand(GroupCommand):
@@ -35,6 +41,7 @@ class ExportCommand(GroupCommand):
             None,
             "Include development dependencies. (<warning>Deprecated</warning>)",
         ),
+        option("all-groups", None, "Include all dependency groups"),
         option(
             "with",
             None,
@@ -66,6 +73,7 @@ class ExportCommand(GroupCommand):
             flag=False,
             multiple=True,
         ),
+        option("all-extras", None, "Include all sets of extra dependencies."),
         option("with-credentials", None, "Include credentials for extra indices."),
     ]
 
@@ -108,20 +116,49 @@ class ExportCommand(GroupCommand):
                 "</warning>"
             )
 
-        # Checking extras
-        extras = {
-            canonicalize_name(extra)
-            for extra_opt in self.option("extras")
-            for extra in extra_opt.split()
-        }
-        invalid_extras = extras - self.poetry.package.extras.keys()
-        if invalid_extras:
-            raise ValueError(
-                f"Extra [{', '.join(sorted(invalid_extras))}] is not specified."
+        if self.option("extras") and self.option("all-extras"):
+            self.line_error(
+                "<error>You cannot specify explicit"
+                " `<fg=yellow;options=bold>--extras</>` while exporting"
+                " using `<fg=yellow;options=bold>--all-extras</>`.</error>"
             )
+            return 1
+
+        extras: Iterable[NormalizedName]
+        if self.option("all-extras"):
+            extras = self.poetry.package.extras.keys()
+        else:
+            extras = {
+                canonicalize_name(extra)
+                for extra_opt in self.option("extras")
+                for extra in extra_opt.split()
+            }
+            invalid_extras = extras - self.poetry.package.extras.keys()
+            if invalid_extras:
+                raise ValueError(
+                    f"Extra [{', '.join(sorted(invalid_extras))}] is not specified."
+                )
+
+        if (
+            self.option("with") or self.option("without") or self.option("only")
+        ) and self.option("all-groups"):
+            self.line_error(
+                "<error>You cannot specify explicit"
+                " `<fg=yellow;options=bold>--with</>`, "
+                "`<fg=yellow;options=bold>--without</>`, "
+                "or `<fg=yellow;options=bold>--only</>` "
+                "while exporting using `<fg=yellow;options=bold>--all-groups</>`.</error>"
+            )
+            return 1
+
+        groups = (
+            self.poetry.package.dependency_group_names(include_optional=True)
+            if self.option("all-groups")
+            else self.activated_groups
+        )
 
         exporter = Exporter(self.poetry, self.io)
-        exporter.only_groups(list(self.activated_groups))
+        exporter.only_groups(list(groups))
         exporter.with_extras(list(extras))
         exporter.with_hashes(not self.option("without-hashes"))
         exporter.with_credentials(self.option("with-credentials"))