+++ tests/types.py
@@ -20,8 +20,7 @@ class CommandTesterFactory(Protocol):
         installer: Installer | None = None,
         executor: Executor | None = None,
         environment: Env | None = None,
-    ) -> CommandTester:
-        ...
+    ) -> CommandTester: ...
 
 
 class ProjectFactory(Protocol):
@@ -33,5 +32,4 @@ class ProjectFactory(Protocol):
         pyproject_content: str | None = None,
         poetry_lock_content: str | None = None,
         install_deps: bool = True,
-    ) -> Poetry:
-        ...
+    ) -> Poetry: ...